% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manipulacao.R
\name{gerarMaterial}
\alias{gerarMaterial}
\title{Gera o meterial}
\usage{
gerarMaterial(audioOnly = T, modoDev = F)
}
\arguments{
\item{audioOnly}{= T boolean}

\item{modoDev}{= F boolean}
}
\value{
Material gerado
}
\description{
Gera o meterial
}
\examples{
gerarMaterial(T, T)
gerarMaterial(T, F)
}

